refactor: 코드리뷰반영하여 패키지 정리
