+++ src/main/java/carracing/model/RandomMovableStrategy.java
@@ -2,8 +2,6 @@ package carracing.model;
 
 import java.util.Random;
 
-import carracing.GameRule;
-
 public class RandomMovableStrategy implements MovableStrategy<Integer> {
     private static final Random random = new Random();
