+++ src/test/java/carracing/controller/GameRuleTest.java
@@ -1,8 +1,10 @@
-package carracing;
+package carracing.controller;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import carracing.model.GameRule;
 
 public class GameRuleTest {
