+++ src/test/java/carracing/model/CarTest.java
@@ -5,8 +5,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import carracing.GameRule;
-
 public class CarTest {
 
     @DisplayName("4미만은 멈춘다")
